requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
langchain==0.3.19
langserve==0.3.1
chromadb==0.6.3
requests==2.32.3
python-multipart==0.0.20
jinja2==3.1.5
pydantic==2.10.6
pydantic-settings==2.7.1
PyPDF2==3.0.1
python-docx==1.1.2
pytest==8.3.4
httpx==0.28.1
